requirements/test.txt
# Test dependencies go here.
-r base.txt

flake8==3.7.9
django-test-plus==1.4.0
factory-boy==2.12.0
coverage==5.0.3
django-coverage-plugin==1.8.0

# pytest
pytest-django==3.8.0
pytest-sugar==0.9.2

# Selenium for UI testing
selenium==3.141.0

# BeautifulSoup for HTML testing
beautifulsoup4==4.8.2

# Black for formatting
black==19.10b0
